refactor(routes tests): making a feeling spec and removing it from user
